Update graph to allow checkbox selection by region
Following instructions in:
http://isakiko.github.io/D3-visualising-data/09-d3exit.html

Add region checkboxes to index.html
Add selection logic to update graph on checkbox click.

Verified the code against the example:
http://isakiko.github.io/D3-visualising-data/code/index09.html

It is nearly identical but suffers from the same rendering issues
when page is reloaded and not all checkboxes are active.  Force
reload works around the issue.